feat: v6.1.0 release

+++ console/repositories/team_repo.py
@@ -6,7 +6,7 @@ from console.exception.main import ServiceHandleException
 from console.models.main import RegionConfig, TeamGitlabInfo, TeamRegistryAuth
 from console.repositories.base import BaseConnection
 from django.db.models import Q
-from www.models.main import (PermRelTenant, TenantEnterprise, TenantRegionInfo, Tenants, Users)
+from www.models.main import (PermRelTenant, TenantEnterprise, TenantRegionInfo, Tenants, Users, TeamInvitation)
 from www.utils.crypt import make_tenant_id
 
 logger = logging.getLogger("default")
@@ -348,8 +348,11 @@ class TeamGitlabRepo(object):
 
 
 class TeamRegistryAuthRepo(object):
-    def list_by_team_id(self, tenant_id, region_name):
-        return TeamRegistryAuth.objects.filter(tenant_id=tenant_id, region_name=region_name)
+    def list_by_team_id(self, tenant_id, region_name, user_id):
+        return TeamRegistryAuth.objects.filter(tenant_id=tenant_id, region_name=region_name, user_id=user_id)
+
+    def check_exist_registry_auth(self, secret_id, user_id):
+        return TeamRegistryAuth.objects.filter(secret_id=secret_id, user_id=user_id)
 
     def create_team_registry_auth(self, **params):
         return TeamRegistryAuth.objects.create(**params)
@@ -358,8 +361,8 @@ class TeamRegistryAuthRepo(object):
         return TeamRegistryAuth.objects.filter(
             tenant_id=tenant_id, region_name=region_name, secret_id=secret_id).update(**params)
 
-    def delete_team_registry_auth(self, tenant_id, region_name, secret_id):
-        return TeamRegistryAuth.objects.filter(tenant_id=tenant_id, region_name=region_name, secret_id=secret_id).delete()
+    def delete_team_registry_auth(self, tenant_id, region_name, secret_id, user_id):
+        return TeamRegistryAuth.objects.filter(tenant_id=tenant_id, region_name=region_name, secret_id=secret_id, user_id=user_id).delete()
 
     def get_by_secret_id(self, secret_id):
         return TeamRegistryAuth.objects.filter(secret_id=secret_id)
@@ -368,6 +371,33 @@ class TeamRegistryAuthRepo(object):
         return TeamRegistryAuth.objects.filter(tenant_id=tenant_id, region_name=region_name, domain=domain)
 
 
+class TeamInvitationRepo(object):
+    def list_by_user_id(self, user_id):
+        """获取用户收到的所有邀请"""
+        return TeamInvitation.objects.filter(user_id=user_id)
+    
+    def get_invitation_by_id(self, invitation_id):
+        """通过ID获取邀请信息"""
+        return TeamInvitation.objects.filter(invitation_id=invitation_id).first()
+    
+    def list_by_team_id(self, team_id):
+        """获取团队所有邀请记录"""
+        return TeamInvitation.objects.filter(tenant_id=team_id)
+    
+    def create_invitation(self, **params):
+        """创建新的团队邀请"""
+        return TeamInvitation.objects.create(**params)
+    
+    def delete_invitation(self, invitation_id):
+        """删除团队邀请"""
+        return TeamInvitation.objects.filter(invitation_id=invitation_id).delete()
+    
+    def update_invitation(self, invitation_id, **params):
+        """更新邀请信息"""
+        return TeamInvitation.objects.filter(invitation_id=invitation_id).update(**params)
+
+
 team_repo = TeamRepo()
 team_gitlab_repo = TeamGitlabRepo()
 team_registry_auth_repo = TeamRegistryAuthRepo()
+team_invitation_repo = TeamInvitationRepo()
